CORE-8621 Move metadata cells from dialog to cells package
