+++ de-lib/src/main/java/org/iplantc/de/diskResource/client/views/metadata/cells/DownloadTemplateCell.java
@@ -1,10 +1,11 @@
-package org.iplantc.de.diskResource.client.views.metadata.dialogs;
+package org.iplantc.de.diskResource.client.views.metadata.cells;
+
+import static com.google.gwt.dom.client.BrowserEvents.CLICK;
 
 import org.iplantc.de.client.events.EventBus;
 import org.iplantc.de.client.models.diskResources.MetadataTemplateInfo;
+import org.iplantc.de.diskResource.client.MetadataView;
 import org.iplantc.de.diskResource.client.events.TemplateDownloadEvent;
-
-import static com.google.gwt.dom.client.BrowserEvents.CLICK;
 
 import com.google.gwt.cell.client.AbstractCell;
 import com.google.gwt.cell.client.Cell;
@@ -15,37 +16,25 @@ import com.google.gwt.dom.client.NativeEvent;
 import com.google.gwt.dom.client.Node;
 import com.google.gwt.safehtml.shared.SafeHtmlBuilder;
 import com.google.gwt.user.client.Event;
-import com.google.gwt.user.client.Window;
-import com.google.gwt.user.client.ui.HTML;
-
-import com.sencha.gxt.widget.core.client.Dialog;
 
 /**
- * Created by sriram on 7/7/16.
+ * Created by sriram on 6/27/16.
  */
-public class TemplateNameCell extends AbstractCell<MetadataTemplateInfo> {
+public class DownloadTemplateCell extends AbstractCell<MetadataTemplateInfo> {
 
-    public interface TemplateNameCellAppearance {
+    public interface DownloadTemplateCellAppearance {
         void render(SafeHtmlBuilder sb,
-                    MetadataTemplateInfo value);
-
-        String description();
-
-        String background();
+                    String debugId);
+        String download();
     }
 
+    private final DownloadTemplateCellAppearance appearance;
 
-    private TemplateNameCellAppearance appearance;
-
-    public TemplateNameCell() {
-       super(CLICK);
-       appearance = GWT.create(TemplateNameCellAppearance.class);
+    public DownloadTemplateCell() {
+        super(CLICK);
+        appearance = GWT.create(DownloadTemplateCellAppearance.class);
     }
 
-    @Override
-    public void render(Context context, MetadataTemplateInfo value, SafeHtmlBuilder sb) {
-        appearance.render(sb,value);
-    }
 
     @Override
     public void onBrowserEvent(Cell.Context context,
@@ -53,12 +42,11 @@ public class TemplateNameCell extends AbstractCell<MetadataTemplateInfo> {
                                MetadataTemplateInfo value,
                                NativeEvent event,
                                ValueUpdater<MetadataTemplateInfo> valueUpdater) {
-        Element eventTarget = Element.as(event.getEventTarget());
-        if ((value == null) || !parent.isOrHasChild(eventTarget)) {
+        if (value == null) {
             return;
         }
 
-
+        Element eventTarget = Element.as(event.getEventTarget());
         Element child = findAppNameElement(parent);
         if (child != null && child.isOrHasChild(eventTarget)) {
             switch (Event.as(event).getTypeInt()) {
@@ -72,13 +60,14 @@ public class TemplateNameCell extends AbstractCell<MetadataTemplateInfo> {
         }
     }
 
+
     private Element findAppNameElement(Element parent) {
         for (int i = 0; i < parent.getChildCount(); i++) {
             Node childNode = parent.getChild(i);
 
             if (Element.is(childNode)) {
                 Element child = Element.as(childNode);
-                if (child.getAttribute("name").equalsIgnoreCase(appearance.description())) { //$NON-NLS-1$
+                if (child.getAttribute("name").equalsIgnoreCase(appearance.download())) { //$NON-NLS-1$
                     return child;
                 }
             }
@@ -87,13 +76,11 @@ public class TemplateNameCell extends AbstractCell<MetadataTemplateInfo> {
     }
 
     private void doOnClick(Element eventTarget, MetadataTemplateInfo value) {
-        Dialog d = new Dialog();
-        d.setSize("500","100");
-        d.setHideOnButtonClick(true);
-        d.setHeading(appearance.description());
-        HTML desc = new HTML(value.getDescription());
-        desc.setStylePrimaryName(appearance.background());
-        d.add(desc);
-        d.show();
+        EventBus.getInstance().fireEvent(new TemplateDownloadEvent(value.getId()));
+    }
+
+    @Override
+    public void render(Cell.Context context, MetadataTemplateInfo value, SafeHtmlBuilder sb) {
+         appearance.render(sb,value.getId());
     }
 }
